+++ ChangeLog.md
@@ -2,9 +2,9 @@
 
 This project's release branch is `master`. This log is written from the perspective of the release branch: when changes hit `master`, they are considered released, and the date should reflect that release.
 
-## Unreleased
+## 2020-04-28
 
-* Bump Obelisk to version 0.4 which includes a bump to reflex-platform and nixpkgs 19.09
+* Bump Obelisk to version 0.8 which includes a bump to reflex-platform and nixpkgs 19.09
 
 ## 2020-01-10